* Kill off more unused utilities and relocate private method _show_warning to above its usage. * Move these utils from internet since they are shell related and remove rm_tree as it pipenv uses the version from vistir.
