+++ pipenv/utils/internet.py
@@ -1,17 +1,8 @@
-import errno
-import os
 import re
-import shutil
-import stat
-import sys
-import warnings
-from contextlib import contextmanager
 from urllib.parse import urlparse
 from urllib3 import util as urllib3_util
 
 from pipenv.vendor import parse
-from pipenv.vendor.vistir.compat import ResourceWarning
-
 
 requests_session = None  # type: ignore
 
@@ -61,68 +52,6 @@ def download_file(url, filename, max_retries=1):
 
     with open(filename, "wb") as f:
         f.write(r.content)
-
-
-@contextmanager
-def temp_path():
-    """Allow the ability to set os.environ temporarily"""
-    path = [p for p in sys.path]
-    try:
-        yield
-    finally:
-        sys.path = [p for p in path]
-
-
-def is_readonly_path(fn):
-    """Check if a provided path exists and is readonly.
-
-    Permissions check is `bool(path.stat & stat.S_IREAD)` or `not os.access(path, os.W_OK)`
-    """
-    if os.path.exists(fn):
-        return (os.stat(fn).st_mode & stat.S_IREAD) or not os.access(fn, os.W_OK)
-
-    return False
-
-
-def set_write_bit(fn):
-    if isinstance(fn, str) and not os.path.exists(fn):
-        return
-    os.chmod(fn, stat.S_IWRITE | stat.S_IWUSR | stat.S_IRUSR)
-    return
-
-
-def rmtree(directory, ignore_errors=False):
-    shutil.rmtree(
-        directory, ignore_errors=ignore_errors, onerror=handle_remove_readonly
-    )
-
-
-def handle_remove_readonly(func, path, exc):
-    """Error handler for shutil.rmtree.
-
-    Windows source repo folders are read-only by default, so this error handler
-    attempts to set them as writeable and then proceed with deletion."""
-    # Check for read-only attribute
-    default_warning_message = (
-        "Unable to remove file due to permissions restriction: {!r}"
-    )
-    # split the initial exception out into its type, exception, and traceback
-    exc_type, exc_exception, exc_tb = exc
-    if is_readonly_path(path):
-        # Apply write permission and call original function
-        set_write_bit(path)
-        try:
-            func(path)
-        except OSError as e:
-            if e.errno in [errno.EACCES, errno.EPERM]:
-                warnings.warn(default_warning_message.format(path), ResourceWarning)
-                return
-
-    if exc_exception.errno in [errno.EACCES, errno.EPERM]:
-        warnings.warn(default_warning_message.format(path), ResourceWarning)
-        return
-
-    raise exc
 
 
 def get_host_and_port(url):
