+++ pipenv/utils/processes.py
@@ -1,17 +1,14 @@
-import contextlib
 import os
-import signal
 import subprocess
 
 import crayons
 from click import echo as click_echo
-from pipenv.vendor.vistir import run
 
 from pipenv.exceptions import PipenvCmdError
 from pipenv import environments
 
 if environments.MYPY_RUNNING:
-    from typing import Any, Dict, List, Optional, Text, Tuple, Union
+    from typing import Tuple
 
 
 def run_command(cmd, *args, is_verbose=False, **kwargs):
@@ -47,43 +44,6 @@ def run_command(cmd, *args, is_verbose=False, **kwargs):
     if c.returncode and catch_exceptions:
         raise PipenvCmdError(cmd.cmdify(), c.stdout, c.stderr, c.returncode)
     return c
-
-
-@contextlib.contextmanager
-def interrupt_handled_subprocess(
-    cmd, verbose=False, return_object=True, write_to_stdout=False, combine_stderr=True,
-    block=True, nospin=True, env=None
-):
-    """Given a :class:`subprocess.Popen` instance, wrap it in exception handlers.
-
-    Terminates the subprocess when and if a `SystemExit` or `KeyboardInterrupt` are
-    processed.
-
-    Arguments:
-        :param str cmd: A command to run
-        :param bool verbose: Whether to run with verbose mode enabled, default False
-        :param bool return_object: Whether to return a subprocess instance or a 2-tuple, default True
-        :param bool write_to_stdout: Whether to write directly to stdout, default False
-        :param bool combine_stderr: Whether to combine stdout and stderr, default True
-        :param bool block: Whether the subprocess should be a blocking subprocess, default True
-        :param bool nospin: Whether to suppress the spinner with the subprocess, default True
-        :param Optional[Dict[str, str]] env: A dictionary to merge into the subprocess environment
-        :return: A subprocess, wrapped in exception handlers, as a context manager
-        :rtype: :class:`subprocess.Popen` obj: An instance of a running subprocess
-    """
-    obj = run(
-        cmd, verbose=verbose, return_object=True, write_to_stdout=False,
-        combine_stderr=False, block=True, nospin=True, env=env,
-    )
-    try:
-        yield obj
-    except (SystemExit, KeyboardInterrupt):
-        if os.name == "nt":
-            os.kill(obj.pid, signal.CTRL_BREAK_EVENT)
-        else:
-            os.kill(obj.pid, signal.SIGINT)
-        obj.wait()
-        raise
 
 
 def subprocess_run(
