+++ pipenv/utils/shell.py
@@ -3,14 +3,20 @@ import os
 import re
 import shlex
 import shutil
+import stat
+import sys
+import warnings
 from contextlib import contextmanager
 from functools import lru_cache
 import posixpath
 from pathlib import Path
 
-from .dependencies import is_file
+from .constants import SCHEME_LIST
 from .processes import subprocess_run
 from pipenv import environments
+from pipenv.vendor.vistir.compat import ResourceWarning
+
+
 if environments.MYPY_RUNNING:
     from typing import Text
 
@@ -209,6 +215,21 @@ def get_workon_home():
     return expanded_path
 
 
+def is_file(package):
+    """Determine if a package name is for a File dependency."""
+    if hasattr(package, "keys"):
+        return any(key for key in package.keys() if key in ["file", "path"])
+
+    if os.path.exists(str(package)):
+        return True
+
+    for start in SCHEME_LIST:
+        if str(package).startswith(start):
+            return True
+
+    return False
+
+
 def is_virtual_environment(path):
     """Check if a given path is a virtual environment's root.
 
@@ -348,3 +369,59 @@ def normalize_drive(path):
         return f"{drive.upper()}{tail}"
 
     return path
+
+
+@contextmanager
+def temp_path():
+    """Allow the ability to set os.environ temporarily"""
+    path = [p for p in sys.path]
+    try:
+        yield
+    finally:
+        sys.path = [p for p in path]
+
+
+def is_readonly_path(fn):
+    """Check if a provided path exists and is readonly.
+
+    Permissions check is `bool(path.stat & stat.S_IREAD)` or `not os.access(path, os.W_OK)`
+    """
+    if os.path.exists(fn):
+        return (os.stat(fn).st_mode & stat.S_IREAD) or not os.access(fn, os.W_OK)
+
+    return False
+
+
+def set_write_bit(fn):
+    if isinstance(fn, str) and not os.path.exists(fn):
+        return
+    os.chmod(fn, stat.S_IWRITE | stat.S_IWUSR | stat.S_IRUSR)
+    return
+
+
+def handle_remove_readonly(func, path, exc):
+    """Error handler for shutil.rmtree.
+
+    Windows source repo folders are read-only by default, so this error handler
+    attempts to set them as writeable and then proceed with deletion."""
+    # Check for read-only attribute
+    default_warning_message = (
+        "Unable to remove file due to permissions restriction: {!r}"
+    )
+    # split the initial exception out into its type, exception, and traceback
+    exc_type, exc_exception, exc_tb = exc
+    if is_readonly_path(path):
+        # Apply write permission and call original function
+        set_write_bit(path)
+        try:
+            func(path)
+        except OSError as e:
+            if e.errno in [errno.EACCES, errno.EPERM]:
+                warnings.warn(default_warning_message.format(path), ResourceWarning)
+                return
+
+    if exc_exception.errno in [errno.EACCES, errno.EPERM]:
+        warnings.warn(default_warning_message.format(path), ResourceWarning)
+        return
+
+    raise exc
