+++ tests/unit/test_utils.py
@@ -450,23 +450,3 @@ twine = "*"
         sources = [{}]
         with pytest.raises(PipenvUsageError):
             indexes.prepare_pip_source_args(sources, pip_args=None)
-
-    @pytest.mark.utils
-    def test_parse_python_version(self):
-        ver = dependencies.parse_python_version("Python 3.6.5\n")
-        assert ver == {"major": "3", "minor": "6", "micro": "5"}
-
-    @pytest.mark.utils
-    def test_parse_python_version_suffix(self):
-        ver = dependencies.parse_python_version("Python 3.6.5rc1\n")
-        assert ver == {"major": "3", "minor": "6", "micro": "5"}
-
-    @pytest.mark.utils
-    def test_parse_python_version_270(self):
-        ver = dependencies.parse_python_version("Python 2.7\n")
-        assert ver == {"major": "2", "minor": "7", "micro": "0"}
-
-    @pytest.mark.utils
-    def test_parse_python_version_270_garbage(self):
-        ver = dependencies.parse_python_version("Python 2.7+\n")
-        assert ver == {"major": "2", "minor": "7", "micro": "0"}
